workers: retire workstation agent compatibility execution
