+++ pkg/services/workers/internal/runtime_composition.go
@@ -78,7 +78,6 @@ func NewRuntime(
 		now,
 		processEnvironment,
 		currentWorkingDirectory,
-		nil,
 		contentMaterializer,
 		interpolation,
 		executionPolicy,
