+++ pkg/services/workers/internal/services/runtime_assembly/construction/agent_cutover_test.go
@@ -12,7 +12,6 @@ import (
 	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
 	"github.com/portpowered/infinite-you/pkg/services/providers"
 	"github.com/portpowered/infinite-you/pkg/services/workers"
-	"github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners"
 	workeragentrun "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/workstations/executor/agentrun"
 )
 
@@ -31,7 +30,7 @@ func TestWithRunnerSelectionPreservesRunnerSelectionWiring(t *testing.T) {
 	}
 }
 
-func TestServiceBuildExposesDispatchAndDirect(t *testing.T) {
+func TestServiceBuildDoesNotExposeProviderDirectExecution(t *testing.T) {
 	factory := &cutoverProvidersFake{}
 
 	service := New(
@@ -69,8 +68,8 @@ func TestServiceBuildExposesDispatchAndDirect(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Build() error = %v", err)
 	}
-	if result.Dispatch == nil || result.Direct == nil {
-		t.Fatalf("Build() = %#v, want dispatch and direct executors", result)
+	if result.Dispatch == nil || result.Direct != nil {
+		t.Fatalf("Build() = %#v, want dispatch without a direct provider executor", result)
 	}
 }
 
@@ -111,83 +110,6 @@ func TestServiceBuildWithNilProgressPublisher(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Build() with nil progress publisher error = %v", err)
 	}
-}
-
-func TestAgentRunnerProviderOverrideBuildsRunner(t *testing.T) {
-	service := New(
-		nil,
-		nil,
-		nil,
-		testFactoryDocs,
-		nil,
-		workeragentrun.NewLibraryHarnessAdapter(platformfilesystem.Local{}),
-		testRetryRandom,
-		platformfilesystem.Local{},
-	)
-
-	runner, err := service.agentRunner(
-		runtimefixtures.RuntimeConfigLookupFixture{},
-		&interfaces.FactoryWorkerConfig{Type: interfaces.WorkerTypeModel},
-		logging.NoopLogger{},
-		false,
-		providerStub{},
-		nil,
-	)
-	if err != nil {
-		t.Fatalf("agentRunner() error = %v", err)
-	}
-	if runner == nil {
-		t.Fatal("provider override returned nil runner")
-	}
-}
-
-func TestRegistryRunnerPreservesInferenceRequestContract(t *testing.T) {
-	registry := &captureRunnerRegistry{}
-	runner := registryRunner{registry: registry, identity: runners.InferenceIdentity}
-	request := workers.RunnerExecutionRequest{
-		RunnerID:       workers.RunnerIDCodex,
-		ModelOperation: "transcribe",
-		ModelBindings: []workers.ResolvedModelOperationBinding{{
-			Slot:   "audio",
-			Source: workers.ModelOperationBindingSourceInput,
-		}},
-		RequiredOptionalCapabilities: []workers.RunnerOptionalCapability{
-			workers.RunnerOptionalCapabilityWorkingDirectory,
-		},
-	}
-	if _, err := runner.Execute(t.Context(), request); err != nil {
-		t.Fatalf("registryRunner.Execute() error = %v", err)
-	}
-	if registry.request.Identity != runners.InferenceIdentity {
-		t.Fatalf("registry identity = %q, want inference", registry.request.Identity)
-	}
-	attempt := registry.request.Attempt
-	if attempt.RunnerID != request.RunnerID || attempt.ModelOperation != request.ModelOperation {
-		t.Fatalf("inference selection fields = %#v, want runner %q and operation %q", attempt, request.RunnerID, request.ModelOperation)
-	}
-	if len(attempt.ModelBindings) != 1 || attempt.ModelBindings[0].Slot != request.ModelBindings[0].Slot {
-		t.Fatalf("inference model bindings = %#v, want caller binding", attempt.ModelBindings)
-	}
-	if len(registry.request.RequiredCapabilities) != 1 ||
-		registry.request.RequiredCapabilities[0] != request.RequiredOptionalCapabilities[0] {
-		t.Fatalf("required capabilities = %#v, want caller capabilities", registry.request.RequiredCapabilities)
-	}
-}
-
-type captureRunnerRegistry struct {
-	request runners.ExecuteRequest
-}
-
-func (registry *captureRunnerRegistry) Resolve(runners.ResolutionRequest) (runners.Binding, error) {
-	return runners.Binding{}, nil
-}
-
-func (registry *captureRunnerRegistry) Execute(
-	_ context.Context,
-	request runners.ExecuteRequest,
-) (runners.ExecuteResult, error) {
-	registry.request = request
-	return runners.ExecuteResult{Content: "fixture"}, nil
 }
 
 func TestProvidersRootMapsThrottledFailure(t *testing.T) {
