--- pkg/services/workers/internal/services/workstations/executor/agent.go
@@ -1,822 +1,0 @@
-package executor
-
-import (
-	"context"
-	"errors"
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/portpowered/infinite-you/pkg/platform/jsonvalue"
-	platformrandom "github.com/portpowered/infinite-you/pkg/platform/random"
-	modelprovider "github.com/portpowered/infinite-you/pkg/services/models"
-	"github.com/portpowered/infinite-you/pkg/services/providers"
-	workerrunner "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners/runner"
-
-	workerexecution "github.com/portpowered/infinite-you/pkg/services/workers"
-	workerinvocation "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/workstations/invocation"
-
-	"github.com/portpowered/infinite-you/pkg/platform/logging"
-	interfaces "github.com/portpowered/infinite-you/pkg/services/factory_definitions"
-	"github.com/portpowered/infinite-you/pkg/services/work"
-	runnerinference "github.com/portpowered/infinite-you/pkg/services/workers/internal/services/runners/inference"
-)
-
-// AgentExecutor implements WorkstationRequestExecutor for MODEL_WORKER types.
-// It reads prompt/output inputs resolved by WorkstationExecutor, calls the
-// configured Runner for one inference attempt, and maps the response to a
-// WorkResult. The Runner performs one attempt per call; the executor may retry
-// retryable provider outcomes with caller-owned backoff before returning.
-type AgentExecutor struct {
-	providerExecutor  workerexecution.InvocationExecutor
-	runtimeConfig     interfaces.RuntimeDefinitionLookup
-	decisionEnvelopes interfaces.DecisionEnvelopeService
-	logger            logging.Logger
-	retryConfig       providerRetryConfig
-	clock             func() time.Time
-}
-
-var _ WorkstationRequestExecutor = (*AgentExecutor)(nil)
-
-// NewAgentExecutor creates an AgentExecutor from runtime-loaded config and a Provider.
-func NewAgentExecutor(
-	runtimeConfig interfaces.RuntimeDefinitionLookup,
-	provider providers.Service,
-	logger logging.Logger,
-	clock func() time.Time,
-	decisionEnvelopes ...interfaces.DecisionEnvelopeService,
-) *AgentExecutor {
-	return NewAgentExecutorWithRunner(
-		runtimeConfig,
-		RunnerFromProvider(provider),
-		logger,
-		clock,
-		decisionEnvelopes...,
-	)
-}
-
-// NewAgentExecutorWithRunner creates an AgentExecutor from runtime-loaded config
-// and the shared runner execution contract.
-func NewAgentExecutorWithRunner(
-	runtimeConfig interfaces.RuntimeDefinitionLookup,
-	runner workerexecution.Runner,
-	logger logging.Logger,
-	clock func() time.Time,
-	decisionEnvelopes ...interfaces.DecisionEnvelopeService,
-) *AgentExecutor {
-	return newAgentExecutor(
-		runtimeConfig,
-		workerinvocation.NewRunnerExecutor(runner),
-		logger,
-		clock,
-		firstDecisionEnvelopeService(decisionEnvelopes),
-	)
-}
-
-func newAgentExecutor(
-	runtimeConfig interfaces.RuntimeDefinitionLookup,
-	executor workerexecution.InvocationExecutor,
-	logger logging.Logger,
-	clock func() time.Time,
-	decisionEnvelopes interfaces.DecisionEnvelopeService,
-) *AgentExecutor {
-	return &AgentExecutor{
-		providerExecutor:  executor,
-		runtimeConfig:     runtimeConfig,
-		decisionEnvelopes: decisionEnvelopes,
-		logger:            logging.EnsureLogger(logger),
-		retryConfig:       newProviderRetryConfig(platformrandom.CryptoSource{}),
-		clock:             clock,
-	}
-}
-
-// Execute calls the Provider with one rendered workstation request, parses the
-// response against OutputSchema if present, and returns a WorkResult.
-func (ae *AgentExecutor) Execute(ctx context.Context, request workerexecution.WorkstationExecutionRequest) (workerexecution.WorkResult, error) {
-	start := ae.clock()
-	if request.OutputSchema != "" {
-		if err := validateOutputSchema([]byte(request.OutputSchema)); err != nil {
-			failed := outputSchemaConfigurationFailure(request, err)
-			failed.Metrics = agentWorkMetrics(start, 0, ae.clock)
-			return failed, nil
-		}
-	}
-	workerType := workerTypeForExecutionRequest(request)
-	workerDef, ok := ae.runtimeConfig.Worker(workerType)
-	if !ok {
-		return missingWorkerWorkResult(request.Dispatch, workerType, ae.clock().Sub(start)), nil
-	}
-	workerDef = effectiveWorkerDefinition(request, workerDef)
-
-	workstationDef, _ := ae.runtimeConfig.Workstation(inferenceWorkstationType(request))
-	req := inferenceRequestForExecutionRequest(request, workerDef, workstationDef)
-	diagnostics := workDiagnosticsForInferenceRequest(req)
-
-	resp, retryCount, err := ae.inferWithRetry(ctx, req)
-	if err != nil {
-		return inferenceErrorWorkResult(request.Dispatch, err, diagnostics, retryCount, start, ae.clock), nil
-	}
-	diagnostics = withInferenceResponseDiagnostics(diagnostics, resp, retryCount)
-	if !hasAuthoritativeCompletionEvidence(resp) {
-		return missingCompletionEvidenceWorkResult(request, resp, diagnostics, retryCount, start, ae.clock), nil
-	}
-
-	if ae.decisionEnvelopes != nil &&
-		ae.decisionEnvelopes.UsesGoalRoutingDecisionEnvelope(workstationDef) {
-		return goalRoutingEnvelopeWorkResult(
-			ae.decisionEnvelopes,
-			request,
-			resp,
-			diagnostics,
-			retryCount,
-			start,
-			ae.clock,
-		), nil
-	}
-	if ae.decisionEnvelopes != nil &&
-		ae.decisionEnvelopes.UsesDecisionEnvelopeOutcome(workstationDef) {
-		return decisionEnvelopeWorkResult(
-			ae.decisionEnvelopes,
-			request,
-			resp,
-			diagnostics,
-			retryCount,
-			start,
-			ae.clock,
-		), nil
-	}
-
-	outcome := ae.evaluateOutcome(resp, workerDef)
-	shapedContent, err := ae.canonicalInferenceOutput(resp.Content, workerDef, request.ModelOperation)
-	if err != nil {
-		return workerexecution.WorkResult{
-			DispatchID:   request.Dispatch.DispatchID,
-			TransitionID: request.Dispatch.TransitionID,
-			Outcome:      workerexecution.OutcomeFailed,
-			Output:       resp.Content,
-			Error:        err.Error(),
-			Continuation: cloneContinuation(resp.Continuation),
-			Diagnostics:  diagnostics,
-			Metrics:      agentWorkMetrics(start, retryCount, ae.clock),
-		}, nil
-	}
-	if shapedContent != "" {
-		resp.Content = shapedContent
-	}
-	return ae.workResultForInferenceResponse(request, resp, outcome, diagnostics, retryCount, start)
-}
-
-func hasAuthoritativeCompletionEvidence(resp workerexecution.InferenceResponse) bool {
-	if strings.TrimSpace(resp.Content) == "" {
-		return false
-	}
-	if resp.Outcome != "" {
-		return true
-	}
-	for _, value := range completionEvidenceValues(resp) {
-		switch strings.ToLower(strings.TrimSpace(value)) {
-		case completionEvidenceAgentMessage, completionEvidenceProviderResponse:
-			return true
-		}
-	}
-	return false
-}
-
-const (
-	completionEvidenceAgentMessage     = "agent_message"
-	completionEvidenceProviderResponse = "provider_response"
-	completionEvidenceTaskComplete     = "task_complete"
-	completionEvidenceTurnCompleted    = "turn_completed"
-)
-
-func completionEvidenceValues(resp workerexecution.InferenceResponse) []string {
-	values := make([]string, 0, 2)
-	if resp.Diagnostics != nil {
-		values = append(values, resp.Diagnostics.Metadata[workerexecution.ProviderResponseMetadataCompletionEvidence])
-		if resp.Diagnostics.Provider != nil {
-			values = append(values, resp.Diagnostics.Provider.ResponseMetadata[workerexecution.ProviderResponseMetadataCompletionEvidence])
-		}
-	}
-	return values
-}
-
-func completionValidationFailure(resp workerexecution.InferenceResponse) (string, string) {
-	for _, value := range completionEvidenceValues(resp) {
-		switch strings.ToLower(strings.TrimSpace(value)) {
-		case completionEvidenceTaskComplete, completionEvidenceTurnCompleted:
-			return "provider completion evidence was contradictory", "contradictory_completion"
-		}
-	}
-	if strings.TrimSpace(resp.Content) != "" {
-		return "provider completion evidence was incomplete", "missing_completion_evidence"
-	}
-	return "provider completion evidence was missing", "missing_completion_evidence"
-}
-
-// pkgmaintcheck:ignore-cyclomatic-complexity pre-existing baseline debt recorded 2026-08-08; refactor this code below the maintainability threshold and remove this exemption
-func effectiveWorkerDefinition(request workerexecution.WorkstationExecutionRequest, workerDef *interfaces.FactoryWorkerConfig) *interfaces.FactoryWorkerConfig {
-	if workerDef == nil {
-		return workerDef
-	}
-	hasResolvedSelection := request.Model != "" ||
-		request.ModelProvider != "" ||
-		request.ReasoningEffort != "" ||
-		request.ExecutorProvider != "" ||
-		isInvocationPlaceholder(workerDef.Model) ||
-		isInvocationPlaceholder(workerDef.ModelProvider) ||
-		isInvocationPlaceholder(workerDef.ReasoningEffort) ||
-		isInvocationPlaceholder(workerDef.ExecutorProvider)
-	if !hasResolvedSelection {
-		return workerDef
-	}
-	effective := *workerDef
-	if request.Model != "" || isInvocationPlaceholder(workerDef.Model) {
-		effective.Model = request.Model
-	}
-	if request.ModelProvider != "" || isInvocationPlaceholder(workerDef.ModelProvider) {
-		effective.ModelProvider = request.ModelProvider
-	}
-	if request.ReasoningEffort != "" || isInvocationPlaceholder(workerDef.ReasoningEffort) {
-		effective.ReasoningEffort = request.ReasoningEffort
-	}
-	if request.ExecutorProvider != "" || isInvocationPlaceholder(workerDef.ExecutorProvider) {
-		effective.ExecutorProvider = request.ExecutorProvider
-	}
-	return &effective
-}
-
-func isInvocationPlaceholder(value string) bool {
-	trimmed := strings.TrimSpace(value)
-	return strings.HasPrefix(trimmed, "${") && strings.HasSuffix(trimmed, "}")
-}
-
-func (ae *AgentExecutor) canonicalInferenceOutput(raw string, workerDef *interfaces.FactoryWorkerConfig, operationName string) (string, error) {
-	operationName = strings.TrimSpace(operationName)
-	if operationName == "" || workerDef == nil {
-		return raw, nil
-	}
-	var operation interfaces.ModelOperation
-	var ok bool
-	for _, candidate := range workerDef.Operations {
-		if strings.TrimSpace(candidate.Name) == operationName {
-			operation = candidate
-			ok = true
-			break
-		}
-	}
-	if !ok {
-		return raw, nil
-	}
-	parts, err := runnerinference.WorkContentFromInferenceOutput(raw, operation)
-	if err != nil {
-		return "", fmt.Errorf("inference output shaping failed: %w", err)
-	}
-	if len(parts) == 0 {
-		return raw, nil
-	}
-	return runnerinference.MarshalWorkContentOutput(parts)
-}
-
-func decisionEnvelopeWorkResult(
-	decisionEnvelopes interfaces.DecisionEnvelopeService,
-	request workerexecution.WorkstationExecutionRequest,
-	resp workerexecution.InferenceResponse,
-	diagnostics *workerexecution.WorkDiagnostics,
-	retryCount int,
-	start time.Time,
-	clock func() time.Time,
-) workerexecution.WorkResult {
-	result := decisionEnvelopes.WorkResultFromDecisionEnvelopeJSONOrFailed(
-		request.Dispatch.DispatchID,
-		request.Dispatch.TransitionID,
-		resp.Content,
-	)
-	result.Continuation = cloneContinuation(resp.Continuation)
-	result.Diagnostics = mergeWorkDiagnostics(diagnostics, result.Diagnostics)
-	result.Metrics = agentWorkMetrics(start, retryCount, clock)
-	return result
-}
-
-func goalRoutingEnvelopeWorkResult(
-	decisionEnvelopes interfaces.DecisionEnvelopeService,
-	request workerexecution.WorkstationExecutionRequest,
-	resp workerexecution.InferenceResponse,
-	diagnostics *workerexecution.WorkDiagnostics,
-	retryCount int,
-	start time.Time,
-	clock func() time.Time,
-) workerexecution.WorkResult {
-	result := decisionEnvelopes.
-		WorkResultFromGoalRoutingDecisionEnvelopeJSONOrFailed(
-			request.Dispatch.DispatchID,
-			request.Dispatch.TransitionID,
-			resp.Content,
-		)
-	result.Continuation = cloneContinuation(resp.Continuation)
-	result.Diagnostics = mergeWorkDiagnostics(diagnostics, result.Diagnostics)
-	result.Metrics = agentWorkMetrics(start, retryCount, clock)
-	return result
-}
-
-func firstDecisionEnvelopeService(
-	services []interfaces.DecisionEnvelopeService,
-) interfaces.DecisionEnvelopeService {
-	if len(services) == 0 {
-		return nil
-	}
-	return services[0]
-}
-
-func workerTypeForExecutionRequest(request workerexecution.WorkstationExecutionRequest) string {
-	if request.WorkerType != "" {
-		return request.WorkerType
-	}
-	return request.Dispatch.WorkerType
-}
-
-func missingWorkerWorkResult(dispatch work.WorkDispatch, workerType string, duration time.Duration) workerexecution.WorkResult {
-	return workerexecution.WorkResult{
-		DispatchID:   dispatch.DispatchID,
-		TransitionID: dispatch.TransitionID,
-		Outcome:      workerexecution.OutcomeFailed,
-		Error:        "worker config not found: " + workerType,
-		Metrics:      workerexecution.WorkMetrics{Duration: duration},
-	}
-}
-
-func inferenceErrorWorkResult(dispatch work.WorkDispatch, err error, diagnostics *workerexecution.WorkDiagnostics, retryCount int, start time.Time, clock func() time.Time) workerexecution.WorkResult {
-	providerErr := workerexecution.NormalizeProviderExecutionError(err)
-	failureMetadata := workerexecution.WorkFailureMetadataFromProviderError(providerErr)
-	result := workerexecution.WorkResult{
-		DispatchID:      dispatch.DispatchID,
-		TransitionID:    dispatch.TransitionID,
-		Outcome:         workerexecution.OutcomeFailed,
-		Error:           formatAgentProviderError(err),
-		FailureMetadata: workerexecution.CloneWorkFailureMetadata(failureMetadata),
-		Continuation:    continuationFromError(providerErr),
-		Diagnostics:     mergeWorkDiagnostics(withInferenceErrorDiagnostics(diagnostics, err, retryCount), providerDiagnosticsFromError(providerErr)),
-		Metrics:         agentWorkMetrics(start, retryCount, clock),
-	}
-	if providerErr != nil {
-		result.ProviderFailureKind = providerErr.ProviderFailureKind
-		result.ProviderContinuationFailureKind = providerErr.ProviderContinuationFailureKind
-		result.ProviderContinuationOutcome = providerErr.ProviderContinuationOutcome
-	}
-	return result
-}
-
-func missingCompletionEvidenceWorkResult(
-	request workerexecution.WorkstationExecutionRequest,
-	resp workerexecution.InferenceResponse,
-	diagnostics *workerexecution.WorkDiagnostics,
-	retryCount int,
-	start time.Time,
-	clock func() time.Time,
-) workerexecution.WorkResult {
-	failureMessage, failureClassification := completionValidationFailure(resp)
-	return workerexecution.WorkResult{
-		DispatchID:   request.Dispatch.DispatchID,
-		TransitionID: request.Dispatch.TransitionID,
-		Outcome:      workerexecution.OutcomeFailed,
-		Error:        failureMessage,
-		FailureMetadata: &workerexecution.WorkFailureMetadata{
-			Family: workerexecution.WorkFailureFamilyTerminal,
-			Type:   workerexecution.WorkFailureTypeUnknown,
-		},
-		Continuation: cloneContinuation(resp.Continuation),
-		Diagnostics:  completionValidationDiagnostics(diagnostics, failureClassification),
-		Metrics:      agentWorkMetrics(start, retryCount, clock),
-	}
-}
-
-func (ae *AgentExecutor) workResultForInferenceResponse(request workerexecution.WorkstationExecutionRequest, resp workerexecution.InferenceResponse, outcome workerexecution.WorkOutcome, diagnostics *workerexecution.WorkDiagnostics, retryCount int, start time.Time) (workerexecution.WorkResult, error) {
-	metrics := agentWorkMetrics(start, retryCount, ae.clock)
-	var structuredResult any
-	structuredResultPresent := false
-	if request.OutputSchema != "" {
-		ae.logger.Info("parsing output against schema", "schema_bytes", len(request.OutputSchema))
-		parsed, parseErr := parseOutputAgainstSchema(resp.Content, []byte(request.OutputSchema))
-		if parseErr != nil {
-			return workerexecution.WorkResult{
-				DispatchID:      request.Dispatch.DispatchID,
-				TransitionID:    request.Dispatch.TransitionID,
-				Outcome:         workerexecution.OutcomeFailed,
-				Output:          resp.Content,
-				Error:           "structured output schema violation: " + parseErr.Error(),
-				FailureMetadata: structuredOutputSchemaViolationMetadata(),
-				Continuation:    cloneContinuation(resp.Continuation),
-				Diagnostics:     completionValidationDiagnostics(diagnostics, "missing_required_output"),
-				Metrics:         metrics,
-			}, nil
-		}
-		if outcome == workerexecution.OutcomeAccepted {
-			structuredResult = jsonvalue.Clone(parsed)
-			structuredResultPresent = true
-		}
-		if verdictFailure := structuredOutputFailure(resp.Content, request.OutputContract); verdictFailure != "" {
-			return workerexecution.WorkResult{
-				DispatchID:      request.Dispatch.DispatchID,
-				TransitionID:    request.Dispatch.TransitionID,
-				Outcome:         workerexecution.OutcomeFailed,
-				Output:          resp.Content,
-				Error:           "structured output verdict failed: " + verdictFailure,
-				FailureMetadata: structuredOutputFailureMetadata(),
-				Continuation:    cloneContinuation(resp.Continuation),
-				Diagnostics:     diagnostics,
-				Metrics:         metrics,
-			}, nil
-		}
-	}
-
-	return workerexecution.WorkResult{
-		DispatchID:              request.Dispatch.DispatchID,
-		TransitionID:            request.Dispatch.TransitionID,
-		Outcome:                 outcome,
-		Output:                  resp.Content,
-		StructuredResult:        structuredResult,
-		StructuredResultPresent: structuredResultPresent,
-		Continuation:            cloneContinuation(resp.Continuation),
-		Diagnostics:             diagnostics,
-		Metrics:                 metrics,
-	}, nil
-}
-
-func agentWorkMetrics(start time.Time, retryCount int, clock func() time.Time) workerexecution.WorkMetrics {
-	return workerexecution.WorkMetrics{
-		Duration:   clock().Sub(start),
-		RetryCount: retryCount,
-	}
-}
-
-func inferenceRequestForExecutionRequest(request workerexecution.WorkstationExecutionRequest, workerDef *interfaces.FactoryWorkerConfig, workstationDef *interfaces.FactoryWorkstationConfig) workerexecution.ProviderInferenceRequest {
-	req := workerexecution.ProviderInferenceRequest{
-		Dispatch:                     work.CloneWorkDispatch(request.Dispatch),
-		WorkerType:                   request.WorkerType,
-		WorkstationType:              inferenceWorkstationType(request),
-		RunnerID:                     request.RunnerID,
-		ExecutorProvider:             request.ExecutorProvider,
-		ProjectID:                    request.ProjectID,
-		InputTokens:                  cloneRawInputTokens(request.InputTokens),
-		ModelOperation:               request.ModelOperation,
-		ModelBindings:                workerexecution.CloneResolvedModelOperationBindings(request.ModelBindings),
-		SystemPrompt:                 request.SystemPrompt,
-		UserMessage:                  request.UserMessage,
-		OutputSchema:                 request.OutputSchema,
-		ToolExecutionMode:            workerexecution.RunnerToolExecutionModeRequired,
-		RequiredOptionalCapabilities: requiredRunnerOptionalCapabilities(request),
-		EnvVars:                      cloneEnvVars(request.EnvVars),
-		ProcessEnvironment:           append([]string(nil), request.ProcessEnvironment...),
-		Worktree:                     request.Worktree,
-		WorkingDirectory:             request.WorkingDirectory,
-		Continuation:                 cloneContinuation(request.Continuation),
-	}
-	if workerDef != nil {
-		if executorProvider := strings.TrimSpace(workerDef.ExecutorProvider); executorProvider != "" {
-			req.ExecutorProvider = executorProvider
-		}
-		req.Model = workerDef.Model
-		req.ModelProvider = modelProviderForExecution(workerDef.ModelProvider, workerexecution.ResolvedRunnerSelection{
-			RunnerID: request.RunnerID,
-			Source:   request.RunnerSelectionSource,
-		})
-		req.ReasoningEffort = workerDef.ReasoningEffort
-		req.PrintTimeout = PrintTimeoutFromWorkerTimeout(workerDef.Timeout)
-		req.ModelLocality = workerDef.ModelLocality
-		req.SessionID = workerDef.SessionID
-		if workerDef.SessionID != "" {
-			req.RequiredOptionalCapabilities = append(req.RequiredOptionalCapabilities, workerexecution.RunnerOptionalCapabilitySessionResume)
-		}
-	}
-	if strings.TrimSpace(req.RunnerID) == "" {
-		// Workstation selection is resolved before this request reaches the
-		// runner. Keep a provider-backed fallback for direct package fixtures.
-		req.RunnerID = strings.TrimSpace(req.ModelProvider)
-	}
-	return req
-}
-
-func cloneContinuation(reference *workerexecution.ProviderContinuationRef) *workerexecution.ProviderContinuationRef {
-	if reference == nil {
-		return nil
-	}
-	cloned := reference.Clone()
-	return &cloned
-}
-
-func modelProviderForExecution(workerModelProvider string, selection workerexecution.ResolvedRunnerSelection) string {
-	if selection.Source == workerexecution.RunnerSelectionSourceWorkstation ||
-		selection.Source == workerexecution.RunnerSelectionSourceFactory ||
-		selection.Source == workerexecution.RunnerSelectionSourceLegacyProvider {
-		if provider := modelProviderForRunnerID(selection.RunnerID); provider != "" {
-			return provider
-		}
-		if strings.TrimSpace(selection.RunnerID) != "" {
-			// Open / externally supplied conductor identities keep their canonical ID.
-			return selection.RunnerID
-		}
-	}
-	if workerModelProvider != "" {
-		if provider := modelProviderForRunnerID(workerModelProvider); provider != "" {
-			return provider
-		}
-		return workerModelProvider
-	}
-	return modelProviderForRunnerID(selection.RunnerID)
-}
-
-func modelProviderForRunnerID(runnerID string) string {
-	switch workerrunner.NormalizeRunnerID(runnerID) {
-	case workerexecution.RunnerIDCodex:
-		return string(modelprovider.ProviderCodex)
-	case string(modelprovider.ProviderClaude):
-		return string(modelprovider.ProviderClaude)
-	case workerexecution.RunnerIDAntigravity:
-		return string(modelprovider.ProviderAntigravity)
-	default:
-		return ""
-	}
-}
-
-func inferenceWorkstationType(request workerexecution.WorkstationExecutionRequest) string {
-	if request.WorkstationType != "" {
-		return request.WorkstationType
-	}
-	return request.Dispatch.WorkstationName
-}
-
-func continuationFromError(providerErr *workerexecution.ProviderError) *workerexecution.ProviderContinuationRef {
-	if providerErr == nil {
-		return nil
-	}
-	return cloneContinuation(providerErr.Continuation)
-}
-
-func providerDiagnosticsFromError(providerErr *workerexecution.ProviderError) *workerexecution.WorkDiagnostics {
-	if providerErr == nil {
-		return nil
-	}
-	return providerErr.Diagnostics
-}
-
-func formatAgentProviderError(err error) string {
-	var providerErr *workerexecution.ProviderError
-	if errors.As(err, &providerErr) {
-		message := strings.TrimSpace(providerErr.Message)
-		if providerErr.Type == workerexecution.WorkFailureTypeTimeout && message == "execution timeout" {
-			return message
-		}
-		if message != "" {
-			return fmt.Sprintf("%s: %s", providerErr.Error(), message)
-		}
-		return providerErr.Error()
-	}
-	return "provider error: " + err.Error()
-}
-
-func (ae *AgentExecutor) inferWithRetry(
-	ctx context.Context,
-	req workerexecution.ProviderInferenceRequest,
-) (workerexecution.InferenceResponse, int, error) {
-	logger := logging.EnsureLogger(ae.logger)
-	retryCount := 0
-	if ae.providerExecutor == nil {
-		return workerexecution.InferenceResponse{}, retryCount, workerexecution.NewProviderError(
-			workerexecution.WorkFailureTypeMisconfigured,
-			"provider execution requires a provider",
-			nil,
-		)
-	}
-
-	for {
-		result, err := ae.providerExecutor.Execute(ctx, workerexecution.InvocationInput{
-			Request: req,
-			Attempt: retryCount + 1,
-		})
-		if err == nil {
-			return result.Response, retryCount, nil
-		}
-
-		providerErr := workerexecution.NormalizeProviderExecutionError(err)
-		if providerErr == nil {
-			return workerexecution.InferenceResponse{}, retryCount, err
-		}
-
-		decision := workerexecution.WorkFailureDecisionFromProviderError(providerErr)
-		if !decision.Retryable || retryCount >= ae.retryConfig.maxRetries {
-			return workerexecution.InferenceResponse{}, retryCount, providerErr
-		}
-		continuation := providerErr.Continuation
-		if continuation == nil {
-			continuation = result.Continuation
-		}
-		if continuation != nil && strings.TrimSpace(continuation.ProviderSessionID) != "" {
-			req.SessionID = strings.TrimSpace(continuation.ProviderSessionID)
-			req.RequiredOptionalCapabilities = appendRunnerCapabilityIfMissing(
-				req.RequiredOptionalCapabilities,
-				workerexecution.RunnerOptionalCapabilitySessionResume,
-			)
-		}
-
-		baseDelay := ae.retryConfig.initialBackoff << retryCount
-		jitter, jitterErr := ae.retryConfig.jitter(baseDelay)
-		if jitterErr != nil {
-			return workerexecution.InferenceResponse{}, retryCount, fmt.Errorf("select provider retry jitter: %w", jitterErr)
-		}
-		delay := baseDelay + jitter
-		retryCount++
-
-		logger.Warn("provider inference failed; retrying",
-			WorkLogFields(req.Dispatch.Execution,
-				"model_provider", req.ModelProvider,
-				"model", req.Model,
-				"retry_count", retryCount,
-				"max_retries", ae.retryConfig.maxRetries,
-				"provider_error_type", string(providerErr.Type),
-				"backoff_ms", delay.Milliseconds())...)
-
-		if err := ae.retryConfig.sleep(ctx, delay); err != nil {
-			return workerexecution.InferenceResponse{}, retryCount, err
-		}
-	}
-}
-
-func appendRunnerCapabilityIfMissing(
-	capabilities []workerexecution.RunnerOptionalCapability,
-	capability workerexecution.RunnerOptionalCapability,
-) []workerexecution.RunnerOptionalCapability {
-	for _, existing := range capabilities {
-		if existing == capability {
-			return capabilities
-		}
-	}
-	return append(capabilities, capability)
-}
-
-const (
-	defaultProviderMaxRetries     = 2
-	defaultProviderInitialBackoff = 100 * time.Millisecond
-)
-
-type retrySleepFunc func(context.Context, time.Duration) error
-type retryJitterFunc func(time.Duration) (time.Duration, error)
-
-type providerRetryConfig struct {
-	maxRetries     int
-	initialBackoff time.Duration
-	sleep          retrySleepFunc
-	jitter         retryJitterFunc
-}
-
-func newProviderRetryConfig(random platformrandom.Source) providerRetryConfig {
-	return providerRetryConfig{
-		maxRetries:     defaultProviderMaxRetries,
-		initialBackoff: defaultProviderInitialBackoff,
-		sleep:          sleepWithContext,
-		jitter:         retryJitter(random),
-	}
-}
-
-func sleepWithContext(ctx context.Context, delay time.Duration) error {
-	timer := time.NewTimer(delay)
-	defer timer.Stop()
-
-	select {
-	case <-timer.C:
-		return nil
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
-func retryJitter(random platformrandom.Source) retryJitterFunc {
-	return func(baseDelay time.Duration) (time.Duration, error) {
-		if baseDelay <= 0 {
-			return 0, nil
-		}
-
-		maxJitter := baseDelay / 2
-		if maxJitter <= 0 {
-			return 0, nil
-		}
-		if random == nil {
-			return 0, fmt.Errorf("provider retry random source is required")
-		}
-		value, err := random.Int63n(int64(maxJitter) + 1)
-		if err != nil {
-			return 0, err
-		}
-		return time.Duration(value), nil
-	}
-}
-
-func (ae *AgentExecutor) inferOnce(
-	ctx context.Context,
-	req workerexecution.ProviderInferenceRequest,
-) (workerexecution.InferenceResponse, int, error) {
-	if ae.providerExecutor == nil {
-		return workerexecution.InferenceResponse{}, 0, workerexecution.NewProviderError(
-			workerexecution.WorkFailureTypeMisconfigured,
-			"provider execution requires a provider",
-			nil,
-		)
-	}
-	result, err := ae.providerExecutor.Execute(ctx, workerexecution.InvocationInput{
-		Request: req,
-		Attempt: 1,
-	})
-	if err != nil {
-		return workerexecution.InferenceResponse{}, 0, err
-	}
-	return result.Response, 0, nil
-}
-
-type providerRunnerAdapter struct {
-	executor workerexecution.InvocationExecutor
-}
-
-// RunnerFromProvider adapts the Providers root onto the shared runner
-// execution contract. The adapter is a composition edge; it does not define
-// a provider protocol or retain provider/session state in Workers.
-func RunnerFromProvider(provider providers.Service) workerexecution.Runner {
-	return providerRunnerAdapter{executor: workerinvocation.NewExecutor(provider)}
-}
-
-func (a providerRunnerAdapter) Execute(ctx context.Context, request workerexecution.RunnerExecutionRequest) (workerexecution.RunnerExecutionResult, error) {
-	if a.executor == nil {
-		return workerexecution.RunnerExecutionResult{}, workerexecution.NewProviderError(
-			workerexecution.WorkFailureTypeMisconfigured,
-			"runner requires a provider implementation",
-			nil,
-		)
-	}
-	result, err := a.executor.Execute(ctx, workerexecution.InvocationInput{Request: request, Attempt: 1})
-	return result.Response, err
-}
-
-func requiredRunnerOptionalCapabilities(request workerexecution.WorkstationExecutionRequest) []workerexecution.RunnerOptionalCapability {
-	capabilities := make([]workerexecution.RunnerOptionalCapability, 0, 5)
-	if request.OutputSchema != "" {
-		capabilities = append(capabilities, workerexecution.RunnerOptionalCapabilityStructuredOutput)
-	}
-	if request.WorkingDirectoryAuthored && request.WorkingDirectory != "" {
-		capabilities = append(capabilities, workerexecution.RunnerOptionalCapabilityWorkingDirectory)
-	}
-	if shouldRequireWorktreeRunnerCapability(request) {
-		capabilities = append(capabilities, workerexecution.RunnerOptionalCapabilityWorktree)
-	}
-	for _, token := range cloneInputTokens(request.InputTokens) {
-		if tokenHasImageContent(token) {
-			capabilities = append(capabilities, workerexecution.RunnerOptionalCapabilityImageInput)
-			break
-		}
-	}
-	return capabilities
-}
-
-func shouldRequireWorktreeRunnerCapability(request workerexecution.WorkstationExecutionRequest) bool {
-	if request.Worktree == "" {
-		return false
-	}
-	if request.WorkingDirectory != "" && workerrunner.NormalizeRunnerID(request.RunnerID) == workerexecution.RunnerIDCodex {
-		return false
-	}
-	return true
-}
-
-func tokenHasImageContent(token workerexecution.Token) bool {
-	for _, part := range token.Color.Content {
-		if part.Type == work.WorkContentPartTypeImage {
-			return true
-		}
-	}
-	return false
-}
-
-// evaluateOutcome determines the WorkOutcome based on stop token evaluation.
-// When no stop token is configured, all successful provider responses are ACCEPTED.
-// When a stop token is configured, the output is checked: found → ACCEPTED,
-// <CONTINUE> → CONTINUE, otherwise → REJECTED.
-func (ae *AgentExecutor) evaluateOutcome(resp workerexecution.InferenceResponse, workerDef *interfaces.FactoryWorkerConfig) workerexecution.WorkOutcome {
-	switch resp.Outcome {
-	case workerexecution.OutcomeAccepted,
-		workerexecution.OutcomeContinue,
-		workerexecution.OutcomeRejected,
-		workerexecution.OutcomeFailed:
-		return resp.Outcome
-	}
-	if workerDef.StopToken == "" {
-		ae.logger.Info("no stop token configured; defaulting to ACCEPTED outcome")
-		return workerexecution.OutcomeAccepted
-	}
-	if workerexecution.ContainsStopToken(resp.Content, workerDef.StopToken) {
-		ae.logger.Info("stop token found in output; returning ACCEPTED outcome", "stop_token", workerDef.StopToken)
-		return workerexecution.OutcomeAccepted
-	}
-	if strings.Contains(resp.Content, "<CONTINUE>") {
-		return workerexecution.OutcomeContinue
-	}
-	return workerexecution.OutcomeRejected
-}
-
-// Compile-time check.
-var _ WorkstationRequestExecutor = (*AgentExecutor)(nil)
