+++ pkg/services/workers/internal/services/workstations/executor/test_fixtures_test.go
@@ -1,0 +1,7 @@
+package executor
+
+import "github.com/portpowered/infinite-you/internal/testutil/runtimefixtures"
+
+// staticRuntimeConfig is shared by the retained workstation prompt/route
+// tests. Agent execution fixtures live with the detached runner tests now.
+type staticRuntimeConfig = runtimefixtures.RuntimeConfigLookupFixture
